# daemon/swagger.yaml
swagger: "2.0"
info:
  title: Software Factory API
  version: "1.0"
  description: |
    API for creating and controlling Software Factory tasks. It is loopback-only
    by default. Local mode uses the GET /control mutation token. Remote mode
    requires the configured bearer credential for every operation and disables
    /control.
basePath: /api/v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  DaemonCredential:
    type: apiKey
    name: Authorization
    in: header
    description: Bearer credential configured through SOFTWARE_FACTORY_DAEMON_TOKEN
  MutationToken:
    type: apiKey
    name: X-Software-Factory-Token
    in: header
paths:
  /identity:
    get:
      tags: [System]
      operationId: getIdentity
      summary: Get the stable daemon identity
      responses:
        "200":
          description: Stable identity for this daemon data directory
          schema:
            $ref: "#/definitions/Identity"
        "401":
          $ref: "#/responses/Unauthorized"
...
  /models:
    get:
      tags: [System]
      operationId: getModels
      summary: List available models for a harness
      parameters:
        - in: query
          name: harness
          type: string
          required: false
          description: Harness name (default pi)
      responses:
        "200":
          description: Available models
          schema:
            $ref: "#/definitions/ModelsResponse"
        "503":
          $ref: "#/responses/ServiceUnavailable"
...
  /tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      summary: List tasks
      responses:
        "200":
          description: Tasks ordered newest first
          schema:
            type: array
            items:
              $ref: "#/definitions/Task"
        "500":
          $ref: "#/responses/InternalError"
    post:
      tags: [Tasks]
      operationId: createTask
      summary: Create a task draft
      security:
        - DaemonCredential: []
        - MutationToken: []
      parameters:
        - in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/CreateTaskRequest"
      responses:
        "201":
          description: Task draft created
          schema:
            $ref: "#/definitions/Task"
        "403":
          $ref: "#/responses/Forbidden"
        "422":
          $ref: "#/responses/UnprocessableEntity"
  /tasks/{id}:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: getTask
      summary: Get a task
      responses:
        "200":
          description: Task
          schema:
            $ref: "#/definitions/Task"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: Delete an inactive task and its files
      security:
        - DaemonCredential: []
        - MutationToken: []
      responses:
        "200":
          description: Task deleted
          schema:
            $ref: "#/definitions/DeletedResponse"
        "403":
          $ref: "#/responses/Forbidden"
        "404":
          $ref: "#/responses/NotFound"
        "409":
          $ref: "#/responses/Conflict"
        "500":
          $ref: "#/responses/InternalError"
  /tasks/{id}/sessions:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskSessions
      summary: List a task's sessions
      responses:
        "200":
          description: Sessions ordered oldest first, including the initial session
          schema:
            type: array
            items:
              $ref: "#/definitions/Task"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
    post:
      tags: [Tasks]
      operationId: createTaskSession
      summary: Create another session in a task
      security:
        - DaemonCredential: []
        - MutationToken: []
      parameters:
        - in: body
          name: session
          required: true
          schema:
            $ref: "#/definitions/CreateSessionRequest"
      responses:
        "201":
          description: Session draft created
          schema:
            $ref: "#/definitions/Task"
        "403":
          $ref: "#/responses/Forbidden"
        "404":
          $ref: "#/responses/NotFound"
        "422":
          $ref: "#/responses/UnprocessableEntity"
...
  /tasks/{id}/feedback:
    parameters:
      - $ref: "#/parameters/TaskID"
    post:
      tags: [Tasks]
      operationId: reviseTaskPlan
      summary: Answer Planner questions and revise the plan
      security:
        - DaemonCredential: []
        - MutationToken: []
      parameters:
        - in: body
          name: feedback
          required: true
          schema:
            $ref: "#/definitions/FeedbackRequest"
      responses:
        "202":
          $ref: "#/responses/Accepted"
        "409":
          $ref: "#/responses/Conflict"
        "422":
          $ref: "#/responses/UnprocessableEntity"
  /tasks/{id}/interventions:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskInterventions
      responses:
        "200":
          description: Persisted Task interventions
          schema:
            type: array
            items:
              $ref: "#/definitions/Intervention"
    post:
      tags: [Tasks]
      operationId: commentOnTaskHistory
      summary: Add an idempotent comment to a Task, Attempt, Event, or Artifact
      security:
        - DaemonCredential: []
        - MutationToken: []
      parameters:
        - in: body
          name: intervention
          required: true
          schema:
            $ref: "#/definitions/InterventionRequest"
      responses:
        "202":
          description: Intervention persisted
          schema:
            $ref: "#/definitions/Intervention"
...
  /tasks/{id}/attempts:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskAttempts
      summary: List Task attempts
      responses:
        "200":
          description: Task attempts
          schema:
            type: array
            items:
              $ref: "#/definitions/Phase"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
  /tasks/{id}/events:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskEvents
      summary: List task events after a cursor or read the newest event tail
      parameters:
        - in: query
          name: after
          description: Return events after this cursor. Ignored when tail is set.
          type: integer
          format: int64
          minimum: 0
          default: 0
        - in: query
          name: limit
          description: Maximum events returned for a cursor read.
          type: integer
          minimum: 1
          maximum: 1000
          default: 250
        - in: query
          name: tail
          description: Return the newest events in sequence order.
          type: integer
          minimum: 1
          maximum: 1000
      responses:
        "200":
          description: Event page and next cursor
          schema:
            $ref: "#/definitions/EventsResponse"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
  /tasks/{id}/events/stream:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: streamTaskEvents
      summary: Stream task events using server-sent events
      produces:
        - text/event-stream
      parameters:
        - in: query
          name: after
          type: integer
          format: int64
          minimum: 0
          default: 0
        - in: header
          name: Last-Event-ID
          type: integer
          format: int64
          required: false
      responses:
        "200":
          description: Server-sent event stream; each event uses the SQLite sequence as its id
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            $ref: "#/definitions/APIError"
        "500":
          description: Streaming unavailable
          schema:
            $ref: "#/definitions/APIError"
  /tasks/{id}/results:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskResults
      summary: List agent result envelopes
      responses:
        "200":
          description: Agent result envelopes
          schema:
            type: array
            items:
              $ref: "#/definitions/Envelope"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
  /tasks/{id}/checks:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: listTaskChecks
      summary: List deterministic check results
      responses:
        "200":
          description: Check results
          schema:
            type: array
            items:
              $ref: "#/definitions/Check"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
  /tasks/{id}/diff:
    parameters:
      - $ref: "#/parameters/TaskID"
    get:
      tags: [Tasks]
      operationId: getTaskDiff
      summary: Get changed files and the repository patch
      responses:
        "200":
          description: Task workspace diff
          schema:
            $ref: "#/definitions/Diff"
        "404":
          $ref: "#/responses/NotFound"
        "500":
          $ref: "#/responses/InternalError"
parameters:
  TaskID:
    in: path
    name: id
    type: string
    required: true
    pattern: '^SF-[0-9]{8}-[a-f0-9]{8}$'
    description: Task identifier
responses:
  Unauthorized:
    description: Daemon credential missing or invalid
    schema:
      $ref: "#/definitions/APIError"
  Accepted:
    description: Command accepted
    schema:
      $ref: "#/definitions/AcceptedResponse"
  Forbidden:
    description: Mutation token missing or invalid, or foreign origin rejected
    schema:
      $ref: "#/definitions/APIError"
  NotFound:
    description: Task not found
    schema:
      $ref: "#/definitions/APIError"
  Conflict:
    description: Task state does not allow the operation
    schema:
      $ref: "#/definitions/APIError"
  UnprocessableEntity:
    description: Invalid request, repository, or service configuration
    schema:
      $ref: "#/definitions/APIError"
  ServiceUnavailable:
    description: Pi model catalog unavailable
    schema:
      $ref: "#/definitions/APIError"
  InternalError:
    description: Internal server error
    schema:
      $ref: "#/definitions/APIError"
definitions:
  Identity:
    type: object
    required: [id]
    properties:
      id:
        type: string
        pattern: "^[a-f0-9]{32}$"
  APIError:
    type: object
    required: [code, message]
    properties:
      code:
        type: string
      message:
        type: string
...
  Model:
    type: object
    required: [provider, id]
    properties:
      provider:
        type: string
      id:
        type: string
      context_window:
        type: integer
        minimum: 0
  ModelsResponse:
    type: object
    required: [harness, models]
    properties:
      harness:
        type: string
      models:
        type: array
        items:
          $ref: "#/definitions/Model"
...
  Thinking:
    type: string
    enum: [off, minimal, low, medium, high, xhigh, max]
...
  CreateTaskRequest:
    type: object
    required: [request, repositories]
    properties:
      request:
        type: string
        minLength: 1
      repositories:
        type: array
        minItems: 1
        items:
          $ref: "#/definitions/Repository"
      coding_agent:
        type: string
        enum: [pi, codex]
        description: Coding agent harness for the task (defaults to server config)
      model:
        type: string
        description: Model ID or provider/model pattern applied to all roles
      thinking:
        $ref: "#/definitions/Thinking"
  CreateSessionRequest:
    type: object
    required: [request]
    properties:
      request:
        type: string
        minLength: 1
  Task:
    type: object
    required: [id, request, workspace_path, repositories, state, total_cost, created_at]
    properties:
      id:
        type: string
      parent_task_id:
        type: string
        description: Root task ID for additional sessions; absent on the initial session
      request:
        type: string
      workspace_path:
        type: string
      primary_repository_path:
        type: string
      repositories:
        type: array
        items:
          $ref: "#/definitions/TaskRepository"
      state:
        type: string
        enum: [draft, preparing, planning, awaiting_plan_approval, building, checking, reviewing, completed, blocked, paused, aborted]
      previous_state:
        type: string
      active_phase:
        type: string
      error:
        type: string
      plan_digest:
        type: string
      approval_actor:
        type: string
      approval_at:
        type: string
        format: date-time
      total_cost:
        type: number
        format: double
      coding_agent:
        type: string
      model:
        type: string
      thinking:
        type: string
      created_at:
        type: string
        format: date-time
      started_at:
        type: string
        format: date-time
      ended_at:
        type: string
        format: date-time
  Phase:
    type: object
    required: [id, task_id, name, kind, owner, description, status, sequence, attempt, retries, started_at]
    properties:
      id:
        type: string
      task_id:
        type: string
      name:
        type: string
      kind:
        type: string
      owner:
        type: string
      description:
        type: string
      status:
        type: string
      error:
        type: string
      sequence:
        type: integer
      attempt:
        type: integer
      retries:
        type: integer
      started_at:
        type: string
        format: date-time
      ended_at:
        type: string
        format: date-time
...
  EventsResponse:
    type: object
    required: [events, cursor]
    properties:
      events:
        type: array
        items:
          $ref: "#/definitions/Event"
      cursor:
        type: integer
        format: int64
  Envelope:
    type: object
    required: [id, task_id, phase_id, agent_role, output_type, payload, created_at, valid, attempt]
    properties:
      id:
        type: string
      task_id:
        type: string
      phase_id:
        type: string
      agent_role:
        type: string
        enum: [planner, builder, reviewer]
      output_type:
        type: string
      payload:
        type: string
        description: JSON result envelope encoded as a string
      created_at:
        type: string
        format: date-time
      valid:
        type: boolean
      attempt:
        type: integer
...
  Diff:
    type: object
    required: [repositories]
    properties:
      repositories:
        type: array
        items:
          $ref: "#/definitions/RepositoryDiff"
...
  FeedbackRequest:
    type: object
    required: [feedback]
    properties:
      feedback: {type: string}
      current_plan_digest: {type: string}
  InterventionRequest:
    type: object
    required: [target_type, target_id, message, idempotency_key]
    properties:
      target_type: {type: string, enum: [task, attempt, event, artifact]}
      target_id: {type: string}
      message: {type: string}
      idempotency_key: {type: string}
  Intervention:
    type: object
    required: [id, task_id, target_type, target_id, actor, intent, text, delivery, idempotency_key, created_at]
    properties:
      id: {type: string}
      task_id: {type: string}
      target_type: {type: string}
      target_id: {type: string}
      actor: {type: string}
      intent: {type: string}
      text: {type: string}
      delivery: {type: string}
      idempotency_key: {type: string}
      created_at: {type: string, format: date-time}
  AcceptedResponse:
    type: object
    required: [accepted]
    properties:
      accepted:
        type: boolean
  DeletedResponse:
    type: object
    required: [deleted]
    properties:
      deleted:
        type: boolean